add separate backend process
fork a separate process for handling backend processing from electron's
main process. in order to maintain the current tooling, we just use a
switch in `main/background.js` and use it as entry point for both the
main and the backend process.

main and backend process communicate via node's message-passing
solution. the main process forwards all ipc requests from the renderer
process to the backend process, but this bears some potential issues and
is subject to change.

the backend process doesn't have access to any electron resources,
therefore props such as `app.appData` need to get passed manually to the
backend.